Add validation helper for Label Values/Keys (#164)
* Add validation helper for Label Values/Keys

This implementation is similar to the implementation in the hcloud-go. To give the "user" as much as verbosity as possible i introduced two methods.

- `validate` which is a simple validator that just returns True/False.add-validator-for-labels
- `validate_verbose` which also returns the corresponding error message (like which key/value is not correctly formatted)

I decided against calling one of both function in the other and instead duplicate the code. This is based on my feeling that the code would be more complex when we as a sample call validate_verbose in validate. Also, it is slightly more readable imho, but we can discuss this :)

Closes https://github.com/hetznercloud/hcloud-python/issues/153

Signed-off-by: Lukas Kämmerling <lukas.kaemmerling@hetzner-cloud.de>

* Fix tox

Signed-off-by: Lukas Kämmerling <lukas.kaemmerling@hetzner-cloud.de>

Signed-off-by: Lukas Kämmerling <lukas.kaemmerling@hetzner-cloud.de>