feat: allow returning root_password in servers rebuild (#276)
Fixes #181

Allow returning a full response during the server rebuild by passing an
option to the rebuild method. The main reason for this is to not break
user land and give users time to upgrade. We also don't want to
introduce a new method `rebuild2` or `rebuild_with_full_response` to
keep the API clean.

The deprecation timeline is the following:
- Introduce the `return_response` argument and deprecated the default
behavior (`return_response=False`)
- Wait for the next major release to remove the ability to only return
the action and deprecated the usage of the `return_response` argument.
- Wait for the next+1 major release to fully remove the
`return_response` argument, or simply leave it unused.