fix: do not copy client
If we copy the client, some bound models that where created during the cached session, will still hold  the reference to the cached_client, even outside the scope of the context manager.

By not copying the client, we ensure that the bound model client is not caching request outside the scope of the context.