test: use request_mock for all bound model tests (#536)
Pass `request_mock` directly to the test functions to configure the
mocked HTTP requests. We do not want to update/modify protected
properties (the mock object) within the client class.

Allows us to move the location of the client `request` function (that is
mocked) in the future, without breaking the tests.

A lot of the diff is from search and replaces and breaking the test
arguments onto multiple lines.

This is the first patch of a larger set. The next step is to use the
same technique for the actual API client (not only the bound models).