test: use request_mock for all clients tests (#537)
Similar to #536 but focused on the resources clients.

Pass request_mock directly to the test functions to configure the mocked
HTTP requests. We do not want to update/modify protected properties (the
mock object) within the client class.

Allows us to move the location of the client request function (that is
mocked) in the future, without breaking the tests.

This also starts adding typing information to the tests, which should
open the door for linting our test code as well.

A lot of the diff is from search and replaces and breaking the test
arguments onto multiple lines.

This is the second patch of a larger set.

Based on top of #536