test: generate bound model methods tests (#546)
In our bound models, many of the methods are only proxies to the api
client methods. But our tests are always testing the full stack down to
the HTTP requests. This lead to a lot of duplicate tests for the API
client and the bound models methods.

This change removes the bound models methods tests, and uses a generic
"proxy method test" for all our bound models.